requirements.txt
requests
numpy
pillow
opencv-python
pymupdf
werkzeug
flask
flask-cors